Remove /examples
